Extract CLI module from main.py
Structure:
  src/cli/
    __init__.py    - exports all actions
    console.py     - rich console helpers (ok/fail/dim/header)
    actions.py     - fetch, build, clean, lint, format, hpc
    interactive.py - questionary arrow-key menu

main.py now just 61 lines of argparse wiring.
