+++ src/cli/__init__.py
@@ -1,0 +1,23 @@
+"""CLI module for ANA-P3 project management."""
+
+from .actions import (
+    fetch_mlflow,
+    run_scripts,
+    build_docs,
+    clean_all,
+    ruff_check,
+    ruff_format,
+    hpc_submit,
+)
+from .interactive import interactive
+
+__all__ = [
+    "fetch_mlflow",
+    "run_scripts",
+    "build_docs",
+    "clean_all",
+    "ruff_check",
+    "ruff_format",
+    "hpc_submit",
+    "interactive",
+]
